+++ src/components/OpenAPI/constants.js
@@ -1,0 +1,10 @@
+import { themeOption } from './styles';
+
+export const MENU_CONTENT_CLASS = 'menu-content';
+export const MENU_TITLE_CONTAINER_CLASS = 'menu-title-container';
+
+export const REDOC_OPTIONS = {
+  hideLoading: true,
+  maxDisplayedEnumValues: 5,
+  theme: themeOption,
+};
